# fuse IDetect