There is a problem, that I've found when converting model for NPU. ImplicitA and ImplicitM layers not letting me run yolov7 on Khadas, so i fused them with convolution. There is almost no difference in precision after fusion, results a little bit different after fusion of ImplicitA and Conv due to restricted precision of floating point numbers.
## With fusion
```
Fusing layers... 
RepConv.fuse_repvgg_block
RepConv.fuse_repvgg_block
RepConv.fuse_repvgg_block
IDetect.fuse
Model Summary: 314 layers, 36907898 parameters, 6194944 gradients, 104.5 GFLOPS
 Convert model to Traced-model... 
 traced_script_module saved! 
 model is traced! 

val: Scanning '/usr/src/app/ScaledYOLOv4-yolov4-csp/COCO/valid2017/labels.cache'
               Class      Images      Labels           P           R      mAP@.5
                 all        5000       36781       0.698       0.582       0.625       0.432
Speed: 1.5/1.4/2.9 ms inference/NMS/total per 352x352 image at batch-size 128

Evaluating pycocotools mAP... saving runs/test/yolov7_640_val/best_predictions.json...
NOTE! Installing ujson may make loading annotations faster.
loading annotations into memory...
Done (t=0.38s)
creating index...
index created!
Loading and preparing results...
DONE (t=4.49s)
creating index...
index created!
Running per image evaluation...
Evaluate annotation type *bbox*
DONE (t=56.68s).
Accumulating evaluation results...
DONE (t=10.50s).
 Average Precision  (AP) @[ IoU=0.50:0.95 | area=   all | maxDets=100 ] = 0.44174
 Average Precision  (AP) @[ IoU=0.50      | area=   all | maxDets=100 ] = 0.63143
 Average Precision  (AP) @[ IoU=0.75      | area=   all | maxDets=100 ] = 0.47279
 Average Precision  (AP) @[ IoU=0.50:0.95 | area= small | maxDets=100 ] = 0.23167
 Average Precision  (AP) @[ IoU=0.50:0.95 | area=medium | maxDets=100 ] = 0.48789
 Average Precision  (AP) @[ IoU=0.50:0.95 | area= large | maxDets=100 ] = 0.64010
 Average Recall     (AR) @[ IoU=0.50:0.95 | area=   all | maxDets=  1 ] = 0.35132
 Average Recall     (AR) @[ IoU=0.50:0.95 | area=   all | maxDets= 10 ] = 0.56074
 Average Recall     (AR) @[ IoU=0.50:0.95 | area=   all | maxDets=100 ] = 0.60447
 Average Recall     (AR) @[ IoU=0.50:0.95 | area= small | maxDets=100 ] = 0.39850
 Average Recall     (AR) @[ IoU=0.50:0.95 | area=medium | maxDets=100 ] = 0.66029
 Average Recall     (AR) @[ IoU=0.50:0.95 | area= large | maxDets=100 ] = 0.79839
```
## Without fusion
```
Fusing layers... 
RepConv.fuse_repvgg_block
RepConv.fuse_repvgg_block
RepConv.fuse_repvgg_block
Model Summary: 314 layers, 36907898 parameters, 6194944 gradients, 104.5 GFLOPS
 Convert model to Traced-model... 
 traced_script_module saved! 
 model is traced! 

val: Scanning '/usr/src/app/ScaledYOLOv4-yolov4-csp/COCO/valid2017/labels.cache'
               Class      Images      Labels           P           R      mAP@.5
                 all        5000       36781       0.698       0.582       0.625       0.432
Speed: 1.5/1.4/2.8 ms inference/NMS/total per 352x352 image at batch-size 128

Evaluating pycocotools mAP... saving runs/test/yolov7_640_val2/best_predictions.json...
NOTE! Installing ujson may make loading annotations faster.
loading annotations into memory...
Done (t=0.37s)
creating index...
index created!
Loading and preparing results...
DONE (t=4.67s)
creating index...
index created!
Running per image evaluation...
Evaluate annotation type *bbox*
DONE (t=53.95s).
Accumulating evaluation results...
DONE (t=10.14s).
 Average Precision  (AP) @[ IoU=0.50:0.95 | area=   all | maxDets=100 ] = 0.44178
 Average Precision  (AP) @[ IoU=0.50      | area=   all | maxDets=100 ] = 0.63147
 Average Precision  (AP) @[ IoU=0.75      | area=   all | maxDets=100 ] = 0.47271
 Average Precision  (AP) @[ IoU=0.50:0.95 | area= small | maxDets=100 ] = 0.23174
 Average Precision  (AP) @[ IoU=0.50:0.95 | area=medium | maxDets=100 ] = 0.48796
 Average Precision  (AP) @[ IoU=0.50:0.95 | area= large | maxDets=100 ] = 0.63988
 Average Recall     (AR) @[ IoU=0.50:0.95 | area=   all | maxDets=  1 ] = 0.35123
 Average Recall     (AR) @[ IoU=0.50:0.95 | area=   all | maxDets= 10 ] = 0.56074
 Average Recall     (AR) @[ IoU=0.50:0.95 | area=   all | maxDets=100 ] = 0.60448
 Average Recall     (AR) @[ IoU=0.50:0.95 | area= small | maxDets=100 ] = 0.39854
 Average Recall     (AR) @[ IoU=0.50:0.95 | area=medium | maxDets=100 ] = 0.66046
 Average Recall     (AR) @[ IoU=0.50:0.95 | area= large | maxDets=100 ] = 0.79816
```